pyproject.toml
[project]
name = "driftguard"
version = "0.1.0"
description = "Streaming feature platform with online learning and concept drift detection"
requires-python = ">=3.11"
dependencies = [
    "kafka-python>=2.0",
    "redis>=5.0",
    "river>=0.21",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.23"]
